get rid of onStart
